requirements.txt
python-telegram-bot>=22.0
requests==2.31.0
python-dotenv==1.0.0
supabase>=2.0.0
httpx>=0.27,<0.29
websockets>=12.0
apscheduler==3.10.4
pytz==2024.1
loguru==0.7.2
cryptography==42.0.5
psycopg2-binary==2.9.9
psutil>=5.9.0
